Removed initial get recipe call and moved handle login function to top level. Additionally added conditional rendering to navigate home on successful login.
